old_django/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
Django==1.10.5
django-appconf==1.0.2
django-celery==3.2.1
django-celery-email==1.1.5
django-redis==4.7.0
django-redis-cache==1.7.1
djangorestframework==3.6.2
kombu==3.0.37
psycopg2==2.7.1
pytz==2017.2
redis==2.10.5
uWSGI==2.0.14
pycrypto==2.6.1
django-treebeard==4.1.0
